backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
supabase>=2.4.0
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.1.0
python-dotenv>=0.19.0
pillow>=10.0.0
boto3>=1.28.0
redis>=4.0.0
httpx>=0.24.0
pyjwt>=2.8.0
python-jose[cryptography]>=3.3.0
